Refactor groupShippingRatesByFreeShippingThreshold to separate file.
