+++ js/src/components/flat-shipping-rates-input-cards/minimum-order-card/minimum-order-card.js
@@ -14,30 +14,8 @@ import AppButtonModalTrigger from '.~/components/app-button-modal-trigger';
 import VerticalGapLayout from '.~/components/vertical-gap-layout';
 import MinimumOrderInputControl from './minimum-order-input-control';
 import AddMinimumOrderModal from './add-minimum-order-modal';
+import groupShippingRatesByFreeShippingThreshold from './groupShippingRatesByFreeShippingThreshold';
 import './minimum-order-card.scss';
-
-const groupShippingRatesByFreeShippingThreshold = ( shippingRates ) => {
-	const map = new Map();
-
-	shippingRates.forEach( ( shippingRate ) => {
-		const threshold =
-			shippingRate.options?.free_shipping_threshold === undefined
-				? undefined
-				: Number( shippingRate.options.free_shipping_threshold );
-		const thresholdCurrency = `${ threshold } ${ shippingRate.currency }`;
-
-		const group = map.get( thresholdCurrency ) || {
-			countries: [],
-			threshold,
-			currency: shippingRate.currency,
-		};
-		group.countries.push( shippingRate.country );
-
-		map.set( thresholdCurrency, group );
-	} );
-
-	return Array.from( map.values() );
-};
 
 const MinimumOrderCard = ( props ) => {
 	const { value = [], onChange = noop } = props;
